there's now follow and unfollow
ignore that last commit— `madonctl accounts follow` is not a toggle. you
instead are meant to use the `--unset` flag to unfollow an account. i
should have probably added some checking to make sure that the user is
already followed, but oh well.

also added information to suggest that support for URLs (as there has
been) in both commands.